Sort more intercept tests into smaller modules
