+++ tests/test_intercept.js
@@ -1,7 +1,5 @@
 'use strict'
 
-const path = require('path')
-const nock = require('..')
 const url = require('url')
 const http = require('http')
 const https = require('https')
@@ -10,17 +8,9 @@ const mikealRequest = require('request')
 const superagent = require('superagent')
 const needle = require('needle')
 const restify = require('restify-clients')
-const domain = require('domain')
 const hyperquest = require('hyperquest')
-const async = require('async')
 const got = require('got')
-const lolex = require('lolex')
-const proxyquire = require('proxyquire').noPreserveCache()
-const debug = require('debug')
-const sinon = require('sinon')
-
-const textFile = path.join(__dirname, '..', 'assets', 'reply_file_1.txt')
-const binaryFile = path.join(__dirname, '..', 'assets', 'reply_file_2.txt.gz')
+const nock = require('..')
 
 nock.enableNetConnect()
 
@@ -41,179 +31,9 @@ test('invalid or missing method parameter throws an exception', t => {
   t.end()
 })
 
-test('double activation throws exception', t => {
-  nock.restore()
-  t.false(nock.isActive())
-
-  nock.activate()
-  t.true(nock.isActive())
-
-  t.throws(() => nock.activate(), { message: 'Nock already active' })
-
-  t.true(nock.isActive())
-
-  t.end()
-})
-
-test("when request's content-type is json: reply callback's requestBody should automatically parse to JSON", async t => {
-  const requestBodyFixture = {
-    id: 1,
-    name: 'bob',
-  }
-
-  const scope = nock('http://service')
-    .post('/endpoint')
-    .reply(200, (uri, requestBody) => {
-      t.deepEqual(requestBody, requestBodyFixture)
-      return 'overwrite'
-    })
-
-  const { body } = await got.post('http://service/endpoint', {
-    headers: { 'Content-Type': 'application/json' },
-    body: JSON.stringify(requestBodyFixture),
-  })
-
-  t.equal(body, 'overwrite')
-  scope.done()
-})
-
 test("when the path doesn't include a leading slash it raises an error", function(t) {
   t.plan(1)
   t.throws(() => nock('http://example.test').get('no-leading-slash'))
-})
-
-test("when request has no content-type header: reply callback's requestBody should not automatically parse to JSON", async t => {
-  const requestBodyFixture = {
-    id: 1,
-    name: 'bob',
-  }
-
-  const scope = nock('http://service')
-    .post('/endpoint')
-    .reply(200, (uri, requestBody) => {
-      t.deepEqual(requestBody, JSON.stringify(requestBodyFixture))
-      return 'overwrite'
-    })
-
-  const { body } = await got.post('http://service/endpoint', {
-    body: JSON.stringify(requestBodyFixture),
-  })
-
-  t.equal(body, 'overwrite')
-  scope.done()
-})
-
-test('reply can take a callback', async t => {
-  const scope = nock('http://example.com')
-    .get('/')
-    .reply(200, (path, requestBody, callback) => callback(null, 'Hello World!'))
-
-  const response = await got('http://example.com/', {
-    encoding: null,
-  })
-
-  scope.done()
-  t.type(response.body, Buffer)
-  t.equal(response.body.toString('utf8'), 'Hello World!')
-})
-
-test('reply should send correct statusCode with array-notation and without body', async t => {
-  t.plan(1)
-
-  const expectedStatusCode = 202
-
-  const scope = nock('http://example.com')
-    .get('/')
-    .reply((path, requestBody) => [expectedStatusCode])
-
-  const { statusCode } = await got('http://example.com/')
-
-  t.equal(statusCode, expectedStatusCode)
-  scope.done()
-})
-
-test('reply takes a callback for status code', async t => {
-  const expectedStatusCode = 202
-  const responseBody = 'Hello, world!'
-  const headers = {
-    'X-Custom-Header': 'abcdef',
-  }
-
-  const scope = nock('http://example.com')
-    .get('/')
-    .reply((path, requestBody, cb) => {
-      setTimeout(() => cb(null, [expectedStatusCode, responseBody, headers]), 1)
-    })
-
-  const response = await got('http://example.com/')
-
-  t.equal(response.statusCode, expectedStatusCode, 'sends status code')
-  t.deepEqual(response.headers, headers, 'sends headers')
-  t.equal(response.body, responseBody, 'sends request body')
-  scope.done()
-})
-
-test('reply should throw on error on the callback', t => {
-  let dataCalled = false
-
-  const scope = nock('http://example.com')
-    .get('/')
-    .reply(500, (path, requestBody, callback) =>
-      callback(new Error('Database failed'))
-    )
-
-  // TODO When this request is converted to `got`, it causes the request not
-  // to match.
-  const req = http.request(
-    {
-      host: 'example.com',
-      path: '/',
-      port: 80,
-    },
-    res => {
-      t.equal(res.statusCode, 500, 'Status code is 500')
-
-      res.on('data', data => {
-        dataCalled = true
-        t.ok(data instanceof Buffer, 'data should be buffer')
-        t.ok(
-          data.toString().indexOf('Error: Database failed') === 0,
-          'response should match'
-        )
-      })
-
-      res.on('end', () => {
-        t.ok(dataCalled, 'data handler was called')
-        scope.done()
-        t.end()
-      })
-    }
-  )
-
-  req.end()
-})
-
-test('match debugging works', async t => {
-  const log = sinon.stub(debug, 'log')
-  debug.enable('nock.interceptor')
-  t.once('end', () => {
-    sinon.restore()
-    debug.disable('nock.interceptor')
-  })
-
-  nock('http://example.test')
-    .post('/deep/link')
-    .reply(200, 'Hello World!')
-
-  const exampleBody = 'Hello yourself!'
-  await got.post('http://example.test/deep/link', { body: exampleBody })
-
-  t.ok(log.calledOnce)
-  t.equal(
-    JSON.parse(log.getCall(0).args[1]).href,
-    'http://example.test/deep/link'
-  )
-  t.equal(JSON.parse(log.getCall(0).args[2]), exampleBody)
 })
 
 test('get gets mocked', async t => {
@@ -309,108 +129,6 @@ test('post, lowercase', t => {
   req.end()
 })
 
-test('get with reply callback', async t => {
-  const scope = nock('http://example.com')
-    .get('/')
-    .reply(200, () => 'OK!')
-
-  const { body } = await got('http://example.com')
-  t.equal(body, 'OK!')
-  scope.done()
-})
-
-test('get to different subdomain with reply callback and filtering scope', async t => {
-  // We scope for www.example.com but through scope filtering we will accept
-  // any <subdomain>.example.com.
-  const scope = nock('http://example.test', {
-    filteringScope: scope => /^http:\/\/.*\.example/.test(scope),
-  })
-    .get('/')
-    .reply(200, () => 'OK!')
-
-  const { body } = await got('http://a.example.test')
-  t.equal(body, 'OK!')
-  scope.done()
-})
-
-test('get with reply callback returning object', async t => {
-  const exampleResponse = { message: 'OK!' }
-
-  const scope = nock('http://example.test')
-    .get('/')
-    .reply(200, () => exampleResponse)
-
-  const { body } = await got('http://example.test')
-  t.equal(body, JSON.stringify(exampleResponse))
-  scope.done()
-})
-
-test('get with reply callback returning array with headers', async t => {
-  const scope = nock('http://example.test')
-    .get('/')
-    .reply(() => [202, 'body', { 'x-key': 'value', 'x-key-2': 'value 2' }])
-
-  const { headers, rawHeaders } = await got('http://example.test/')
-
-  t.deepEqual(headers, {
-    'x-key': 'value',
-    'x-key-2': 'value 2',
-  })
-  t.deepEqual(rawHeaders, ['x-key', 'value', 'x-key-2', 'value 2'])
-  scope.done()
-})
-
-// Skipped because https://github.com/nock/nock/issues/1222
-test(
-  'get with reply callback returning default statusCode without body',
-  { skip: true },
-  t => {
-    nock('http://replyheaderland')
-      .get('/')
-      .reply((uri, requestBody) => [401])
-
-    http.get(
-      {
-        host: 'replyheaderland',
-        path: '/',
-        port: 80,
-      },
-      res => {
-        res.setEncoding('utf8')
-        t.equal(res.statusCode, 200)
-        res.on('data', data => {
-          t.equal(data, '[401]')
-          res.once('end', t.end.bind(t))
-        })
-      }
-    )
-  }
-)
-
-test('get with reply callback returning callback without headers', async t => {
-  const scope = nock('http://example.com')
-    .get('/')
-    .reply(() => [401, 'This is a body'])
-
-  await t.rejects(async () => got('http://example.com/'), {
-    statusCode: 401,
-    body: 'This is a body',
-  })
-  scope.done()
-})
-
-test('post with reply callback, uri, and request body', async t => {
-  const input = 'key=val'
-
-  const scope = nock('http://example.com')
-    .post('/echo', input)
-    .reply(200, (uri, body) => ['OK', uri, body].join(' '))
-
-  const { body } = await got('http://example.com/echo', { body: input })
-  t.equal(body, 'OK /echo key=val')
-  scope.done()
-})
-
 test('post with regexp as spec', async t => {
   const input = 'key=val'
 
@@ -485,239 +203,6 @@ test('filteringRequestBody with invalid argument throws expected', t => {
   t.end()
 })
 
-test('isDone', async t => {
-  const scope = nock('http://example.com')
-    .get('/')
-    .reply(200, 'Hello World!')
-
-  t.notOk(scope.isDone(), 'not done when a request is outstanding')
-
-  await got('http://example.com/')
-
-  t.true(scope.isDone(), 'done after request is made')
-  scope.done()
-})
-
-test('headers work', async t => {
-  const scope = nock('http://example.com')
-    .get('/')
-    .reply(200, 'Hello World!', { 'X-My-Headers': 'My Header value' })
-
-  const { headers } = await got('http://example.com/')
-
-  t.equivalent(headers, { 'x-my-headers': 'My Header value' })
-  scope.done()
-})
-
-test('reply headers work with function', async t => {
-  const scope = nock('http://example.com')
-    .get('/')
-    .reply(200, () => 'ABC', { 'X-My-Headers': 'My custom header value' })
-
-  const { headers } = await got('http://example.com/')
-
-  t.equivalent(headers, { 'x-my-headers': 'My custom header value' })
-  scope.done()
-})
-
-test('reply headers as function work', async t => {
-  const scope = nock('http://example.com')
-    .get('/')
-    .reply(200, 'boo!', {
-      'X-My-Headers': (req, res, body) => body.toString(),
-    })
-
-  const { headers, rawHeaders } = await got('http://example.com/')
-
-  t.equivalent(headers, { 'x-my-headers': 'boo!' })
-  t.equivalent(rawHeaders, ['X-My-Headers', 'boo!'])
-  scope.done()
-})
-
-test('reply headers as function are evaluated only once per request', async t => {
-  let counter = 0
-  const scope = nock('http://example.com')
-    .get('/')
-    .reply(200, 'boo!', {
-      'X-My-Headers': (req, res, body) => {
-        ++counter
-        return body.toString()
-      },
-    })
-
-  const { headers, rawHeaders } = await got('http://example.com/')
-
-  t.equivalent(headers, { 'x-my-headers': 'boo!' })
-  t.equivalent(rawHeaders, ['X-My-Headers', 'boo!'])
-  scope.done()
-
-  t.equal(counter, 1)
-})
-
-test('reply headers as function are evaluated on each request', async t => {
-  let counter = 0
-  const scope = nock('http://example.com')
-    .get('/')
-    .times(2)
-    .reply(200, 'boo!', {
-      'X-My-Headers': (req, res, body) => `${++counter}`,
-    })
-
-  const { headers, rawHeaders } = await got('http://example.com/')
-  t.equivalent(headers, { 'x-my-headers': '1' })
-  t.equivalent(rawHeaders, ['X-My-Headers', '1'])
-
-  t.equal(counter, 1)
-
-  const { headers: headers2, rawHeaders: rawHeaders2 } = await got(
-    'http://example.com/'
-  )
-  t.equivalent(headers2, { 'x-my-headers': '2' })
-  t.equivalent(rawHeaders2, ['X-My-Headers', '2'])
-
-  t.equal(counter, 2)
-
-  scope.done()
-})
-
-test('match headers', async t => {
-  const scope = nock('http://example.test')
-    .get('/')
-    .matchHeader('x-my-headers', 'My custom Header value')
-    .reply(200, 'Hello World!')
-
-  const { statusCode, body } = await got('http://example.test/', {
-    headers: { 'X-My-Headers': 'My custom Header value' },
-  })
-
-  t.equal(statusCode, 200)
-  t.equal(body, 'Hello World!')
-  scope.done()
-})
-
-test('multiple match headers', async t => {
-  const scope = nock('http://example.test')
-    .get('/')
-    .matchHeader('x-my-headers', 'My custom Header value')
-    .reply(200, 'Hello World!')
-    .get('/')
-    .matchHeader('x-my-headers', 'other value')
-    .reply(200, 'Hello World other value!')
-
-  const response1 = await got('http://example.test/', {
-    headers: { 'X-My-Headers': 'other value' },
-  })
-
-  t.equal(response1.statusCode, 200)
-  t.equal(response1.body, 'Hello World other value!')
-
-  const response2 = await got('http://example.test/', {
-    headers: { 'X-My-Headers': 'My custom Header value' },
-  })
-
-  t.equal(response2.statusCode, 200)
-  t.equal(response2.body, 'Hello World!')
-
-  scope.done()
-})
-
-test('match headers with regexp', async t => {
-  const scope = nock('http://example.test')
-    .get('/')
-    .matchHeader('x-my-headers', /My He.d.r [0-9.]+/)
-    .reply(200, 'Hello World!')
-
-  const { statusCode, body } = await got('http://example.test/', {
-    headers: { 'X-My-Headers': 'My Header 1.0' },
-  })
-
-  t.equal(statusCode, 200)
-  t.equal(body, 'Hello World!')
-  scope.done()
-})
-
-test('match headers on number with regexp', async t => {
-  const scope = nock('http://example.test')
-    .get('/')
-    .matchHeader('x-my-headers', /\d+/)
-    .reply(200, 'Hello World!')
-
-  const { statusCode, body } = await got('http://example.test/', {
-    headers: { 'X-My-Headers': 123 },
-  })
-
-  t.equal(statusCode, 200)
-  t.equal(body, 'Hello World!')
-  scope.done()
-})
-
-test('match headers with function', async t => {
-  const scope = nock('http://example.test')
-    .get('/')
-    .matchHeader('x-my-headers', val => val > 123)
-    .reply(200, 'Hello World!')
-
-  const { statusCode, body } = await got('http://example.test/', {
-    headers: { 'X-My-Headers': 456 },
-  })
-
-  t.equal(statusCode, 200)
-  t.equal(body, 'Hello World!')
-  scope.done()
-})
-
-test('match all headers', async t => {
-  const scope = nock('http://example.test')
-    .matchHeader('accept', 'application/json')
-    .get('/one')
-    .reply(200, { hello: 'world' })
-    .get('/two')
-    .reply(200, { a: 1, b: 2, c: 3 })
-
-  const response1 = await got('http://example.test/one', {
-    headers: { Accept: 'application/json' },
-  })
-  t.equal(response1.statusCode, 200)
-  t.equal(response1.body, '{"hello":"world"}')
-
-  const response2 = await got('http://example.test/two', {
-    headers: { Accept: 'application/json' },
-  })
-  t.equal(response2.statusCode, 200)
-  t.equal(response2.body, '{"a":1,"b":2,"c":3}')
-
-  scope.done()
-})
-
-test('header manipulation', t => {
-  // This test seems to depend on behavior of the `http` module.
-  const scope = nock('http://example.com')
-    .get('/accounts')
-    .reply(200, { accounts: [{ id: 1, name: 'Joe Blow' }] })
-
-  const req = http.get({ host: 'example.com', path: '/accounts' }, res => {
-    res.on('end', () => {
-      scope.done()
-      t.end()
-    })
-    // Streams start in 'paused' mode and must be started.
-    // See https://nodejs.org/api/stream.html#stream_class_stream_readable
-    res.resume()
-  })
-
-  req.setHeader('X-Custom-Header', 'My Value')
-  t.equal(
-    req.getHeader('X-Custom-Header'),
-    'My Value',
-    'Custom header was not set'
-  )
-
-  req.removeHeader('X-Custom-Header')
-  t.notOk(req.getHeader('X-Custom-Header'), 'Custom header was not removed')
-
-  req.end()
-})
-
 test('head', async t => {
   const scope = nock('http://example.test')
     .head('/')
@@ -776,117 +261,6 @@ test('encoding', async t => {
   t.equal(body, 'SGVsbG8gV29ybGQh', 'response should match base64 encoding')
 
   scope.done()
-})
-
-test('reply with file', async t => {
-  const scope = nock('http://example.test')
-    .get('/')
-    .replyWithFile(200, textFile)
-
-  const { statusCode, body } = await got('http://example.test/')
-
-  t.equal(statusCode, 200)
-  t.equal(body, 'Hello from the file!')
-
-  scope.done()
-})
-
-test('reply with file with headers', async t => {
-  const scope = nock('http://example.test')
-    .get('/')
-    .replyWithFile(200, binaryFile, {
-      'content-encoding': 'gzip',
-    })
-
-  const { statusCode, body } = await got('http://example.test/')
-
-  t.equal(statusCode, 200)
-  t.equal(body.length, 20)
-  scope.done()
-})
-
-test('reply with file with no fs', t => {
-  const nockWithoutFs = proxyquire('../lib/scope', {
-    './interceptor': proxyquire('../lib/interceptor', { fs: null }),
-  })
-
-  t.throws(
-    () =>
-      nockWithoutFs('http://example.test')
-        .get('/')
-        .replyWithFile(200, textFile),
-    {
-      message: 'No fs',
-    }
-  )
-
-  t.end()
-})
-
-test('reply with JSON', async t => {
-  const scope = nock('http://example.test')
-    .get('/')
-    .reply(200, { hello: 'world' })
-
-  const { statusCode, headers, body } = await got('http://example.test/')
-
-  t.equal(statusCode, 200)
-  t.type(headers.date, 'undefined')
-  t.type(headers['content-length'], 'undefined')
-  t.equal(headers['content-type'], 'application/json')
-  t.equal(body, '{"hello":"world"}', 'response should match')
-  scope.done()
-})
-
-test('reply with content-length header', async t => {
-  const scope = nock('http://example.test')
-    .replyContentLength()
-    .get('/')
-    .reply(200, { hello: 'world' })
-
-  const { headers } = await got('http://example.test/')
-
-  t.equal(headers['content-length'], 17)
-  scope.done()
-})
-
-test('reply with explicit date header', async t => {
-  const date = new Date()
-
-  const scope = nock('http://example.test')
-    .replyDate(date)
-    .get('/')
-    .reply(200, { hello: 'world' })
-
-  const { headers } = await got('http://example.test/')
-
-  t.equal(headers.date, date.toUTCString())
-  scope.done()
-})
-
-// async / got version is returning "not ok test unfinished".
-// https://github.com/nock/nock/issues/1305#issuecomment-451701657
-test('reply with implicit date header', t => {
-  const clock = lolex.install()
-  const date = new Date()
-
-  const scope = nock('http://example.test')
-    .replyDate()
-    .get('/')
-    .reply(200)
-
-  mikealRequest.get('http://example.test', (err, resp) => {
-    clock.uninstall()
-
-    if (err) {
-      throw err
-    }
-
-    t.equal(resp.headers.date, date.toUTCString())
-    scope.done()
-
-    t.end()
-  })
 })
 
 test('filter path with function', async t => {
@@ -1224,164 +598,6 @@ test('two scopes with the same request are consumed', async t => {
   scope2.done()
 })
 
-test('JSON encoded replies set the content-type header', async t => {
-  const scope = nock('http://example.test')
-    .get('/')
-    .reply(200, {
-      A: 'b',
-    })
-
-  t.equal(
-    (await got('http://example.test/')).headers['content-type'],
-    'application/json'
-  )
-
-  scope.done()
-})
-
-test('JSON encoded replies does not overwrite existing content-type header', async t => {
-  const scope = nock('http://example.test')
-    .get('/')
-    .reply(
-      200,
-      {
-        A: 'b',
-      },
-      {
-        'Content-Type': 'unicorns',
-      }
-    )
-
-  t.equal(
-    (await got('http://example.test/')).headers['content-type'],
-    'unicorns'
-  )
-
-  scope.done()
-})
-
-test("blank response doesn't have content-type application/json attached to it", async t => {
-  const scope = nock('http://example.test')
-    .get('/')
-    .reply(200)
-
-  t.equal(
-    (await got('http://example.test/')).headers['content-type'],
-    undefined
-  )
-
-  scope.done()
-})
-
-test('unencodable object throws the expected error', t => {
-  const unencodableObject = {
-    toJSON() {
-      throw Error('bad!')
-    },
-  }
-
-  t.throws(
-    () =>
-      nock('http://localhost')
-        .get('/')
-        .reply(200, unencodableObject),
-    {
-      message: 'Error encoding response body into JSON',
-    }
-  )
-
-  t.end()
-})
-
-test('clean all works', t => {
-  nock('http://example.test')
-    .get('/nonexistent')
-    .reply(200)
-
-  http.get({ host: 'example.test', path: '/nonexistent' }, function(res) {
-    t.assert(res.statusCode === 200, 'should mock before cleanup')
-
-    nock.cleanAll()
-
-    http
-      .get({ host: 'example.test', path: '/nonexistent' }, function(res) {
-        res.destroy()
-        t.assert(res.statusCode !== 200, 'should clean up properly')
-        t.end()
-      })
-      .on('error', function(err) {
-        t.end()
-      })
-  })
-})
-
-test('cleanAll should remove pending mocks from all scopes', t => {
-  const scope1 = nock('http://example.test')
-    .get('/somepath')
-    .reply(200, 'hey')
-  t.deepEqual(scope1.pendingMocks(), ['GET http://example.test:80/somepath'])
-  const scope2 = nock('http://example.test')
-    .get('/somepath')
-    .reply(200, 'hey')
-  t.deepEqual(scope2.pendingMocks(), ['GET http://example.test:80/somepath'])
-
-  nock.cleanAll()
-
-  t.deepEqual(scope1.pendingMocks(), [])
-  t.deepEqual(scope2.pendingMocks(), [])
-  t.end()
-})
-
-test('is done works', t => {
-  nock('http://example.test')
-    .get('/nonexistent')
-    .reply(200)
-
-  t.ok(!nock.isDone())
-
-  http.get({ host: 'example.test', path: '/nonexistent' }, function(res) {
-    t.assert(res.statusCode === 200, 'should mock before cleanup')
-    t.ok(nock.isDone())
-    t.end()
-  })
-})
-
-test('pending mocks works', t => {
-  nock('http://example.test')
-    .get('/nonexistent')
-    .reply(200)
-
-  t.deepEqual(nock.pendingMocks(), ['GET http://example.test:80/nonexistent'])
-
-  http.get({ host: 'example.test', path: '/nonexistent' }, function(res) {
-    t.assert(res.statusCode === 200, 'should mock before cleanup')
-    t.deepEqual(nock.pendingMocks(), [])
-    t.end()
-  })
-})
-
-test('activeMocks returns incomplete mocks', t => {
-  nock.cleanAll()
-  nock('http://example.test')
-    .get('/incomplete')
-    .reply(200)
-
-  t.deepEqual(nock.activeMocks(), ['GET http://example.test:80/incomplete'])
-  t.end()
-})
-
-test("activeMocks doesn't return completed mocks", t => {
-  nock.cleanAll()
-  nock('http://example.test')
-    .get('/complete-me')
-    .reply(200)
-
-  http.get({ host: 'example.test', path: '/complete-me' }, function(res) {
-    t.deepEqual(nock.activeMocks(), [])
-    t.end()
-  })
-})
-
 test('username and password works', t => {
   const scope = nock('http://example.test')
     .get('/')
@@ -1569,56 +785,6 @@ test('request has path', t => {
   req.end()
 })
 
-test('(re-)activate after restore', t => {
-  t.plan(7)
-
-  const server = http.createServer((request, response) => {
-    t.pass('server received a request')
-
-    switch (url.parse(request.url).pathname) {
-      case '/':
-        response.writeHead(200)
-        response.write('server served a response')
-        break
-    }
-
-    response.end()
-  })
-
-  server.listen(() => {
-    const scope = nock(`http://localhost:${server.address().port}`)
-      .get('/')
-      .reply(304, 'served from our mock')
-
-    nock.restore()
-    t.false(nock.isActive())
-
-    http.get(`http://localhost:${server.address().port}`, function(res) {
-      res.resume()
-
-      t.is(200, res.statusCode)
-
-      res.on('end', function() {
-        t.ok(!scope.isDone())
-
-        nock.activate()
-        t.true(nock.isActive())
-        http.get(`http://localhost:${server.address().port}`, function(res) {
-          res.resume()
-
-          t.is(304, res.statusCode)
-
-          res.on('end', function() {
-            t.ok(scope.isDone())
-
-            server.close(t.end)
-          })
-        })
-      })
-    })
-  })
-})
-
 test('allow unordered body with json encoding', t => {
   const scope = nock('http://example.test')
     .post('/like-wtf', {
@@ -1710,172 +876,6 @@ test('has a req property on the response', t => {
   req.end()
 })
 
-test('when net connect is disabled, throws the expected error ', async t => {
-  nock.cleanAll()
-  nock.disableNetConnect()
-  t.once('end', () => nock.enableNetConnect())
-
-  try {
-    await got('http://example.test')
-    t.fail('Expected to throw')
-  } catch (err) {
-    t.type(err, 'Error')
-    t.equal(err.message, 'Nock: Disallowed net connect for "example.test:80/"')
-    t.equal(err.code, 'ENETUNREACH')
-    t.ok(err.stack)
-  }
-})
-
-test('enable real HTTP request only for specified domain, via string', t => {
-  t.plan(1)
-
-  const server = http.createServer((request, response) => {
-    t.pass('server received a request')
-    response.writeHead(200)
-    response.end()
-    t.end()
-  })
-  t.once('end', () => server.close())
-
-  nock.enableNetConnect('localhost')
-  t.once('end', () => nock.enableNetConnect())
-
-  server.listen(() =>
-    mikealRequest(`http://localhost:${server.address().port}/`)
-  )
-})
-
-test('disallow request for other domains, via string', t => {
-  nock.enableNetConnect('localhost')
-  t.once('end', () => nock.enableNetConnect())
-
-  http
-    .get('http://www.amazon.com', function(res) {
-      throw 'should not deliver this request'
-    })
-    .on('error', function(err) {
-      t.equal(
-        err.message,
-        'Nock: Disallowed net connect for "www.amazon.com:80/"'
-      )
-      t.end()
-    })
-})
-
-test('enable real HTTP request only for specified domain, via regexp', t => {
-  t.plan(1)
-
-  const server = http.createServer((request, response) => {
-    t.pass('server received a request')
-    response.writeHead(200)
-    response.end()
-    t.end()
-  })
-  t.once('end', () => server.close())
-
-  nock.enableNetConnect(/ocalhos/)
-  t.once('end', () => nock.enableNetConnect())
-
-  server.listen(() =>
-    mikealRequest(`http://localhost:${server.address().port}/`)
-  )
-})
-
-test('disallow request for other domains, via regexp', t => {
-  nock.enableNetConnect(/ocalhos/)
-  t.once('end', () => nock.enableNetConnect())
-
-  http
-    .get('http://www.amazon.com', function(res) {
-      throw 'should not deliver this request'
-    })
-    .on('error', function(err) {
-      t.equal(
-        err.message,
-        'Nock: Disallowed net connect for "www.amazon.com:80/"'
-      )
-      t.end()
-    })
-})
-
-test('repeating once', t => {
-  nock.disableNetConnect()
-
-  nock('http://example.test')
-    .get('/')
-    .once()
-    .reply(200, 'Hello World!')
-
-  http.get('http://example.test', function(res) {
-    t.equal(200, res.statusCode, 'first request')
-    t.end()
-  })
-
-  nock.cleanAll()
-
-  nock.enableNetConnect()
-})
-
-test('repeating twice', t => {
-  nock.disableNetConnect()
-
-  nock('http://example.test')
-    .get('/')
-    .twice()
-    .reply(200, 'Hello World!')
-
-  async.each(
-    [1, 2],
-    function(_, cb) {
-      http.get('http://example.test', function(res) {
-        t.equal(200, res.statusCode)
-        cb()
-      })
-    },
-    t.end.bind(t)
-  )
-})
-
-test('repeating thrice', t => {
-  nock.disableNetConnect()
-
-  nock('http://example.test')
-    .get('/')
-    .thrice()
-    .reply(200, 'Hello World!')
-
-  async.each(
-    [1, 2, 3],
-    function(_, cb) {
-      http.get('http://example.test', function(res) {
-        t.equal(200, res.statusCode)
-        cb()
-      })
-    },
-    t.end.bind(t)
-  )
-})
-
-test('repeating response 4 times', t => {
-  nock.disableNetConnect()
-
-  nock('http://example.test')
-    .get('/')
-    .times(4)
-    .reply(200, 'Hello World!')
-
-  async.each(
-    [1, 2, 3, 4],
-    function(_, cb) {
-      http.get('http://example.test', function(res) {
-        t.equal(200, res.statusCode, 'first request')
-        cb()
-      })
-    },
-    t.end.bind(t)
-  )
-})
-
 test('superagent works', t => {
   const responseText = 'Yay superagent!'
   const headers = { 'Content-Type': 'text/plain' }
@@ -1935,28 +935,6 @@ test('response is an http.IncomingMessage instance', t => {
       }
     )
     .end()
-})
-
-test('resetting nock catastrophically while a request is in progress is handled gracefully', async t => {
-  // While invoking cleanAll() from a nock request handler isn't very
-  // realistic, it's possible that user code under test could crash, causing
-  // before or after hooks to fire, which invoke `nock.cleanAll()`. A little
-  // extreme, though if this does happen, we may as well be graceful about it.
-  function somethingBad() {
-    nock.cleanAll()
-  }
-
-  const scope = nock('http://example.test')
-    .get('/somepath')
-    .reply(200, (uri, requestBody) => {
-      somethingBad()
-      return 'hi'
-    })
-
-  const { body } = await got('http://example.test/somepath')
-
-  t.equal(body, 'hi')
-  scope.done()
 })
 
 test('write callback called', t => {
@@ -2252,194 +1230,6 @@ test('test request timeout option', t => {
   })
 })
 
-test('done fails when specified request header is missing', t => {
-  nock('http://example.test', {
-    reqheaders: {
-      'X-App-Token': 'apptoken',
-      'X-Auth-Token': 'apptoken',
-    },
-  })
-    .post('/resource')
-    .reply(200, { status: 'ok' })
-
-  const d = domain.create()
-
-  d.run(function() {
-    mikealRequest({
-      method: 'POST',
-      uri: 'http://example.test/resource',
-      headers: {
-        'X-App-Token': 'apptoken',
-      },
-    })
-  })
-
-  d.once('error', function(err) {
-    t.ok(err.message.match(/No match/))
-    t.end()
-  })
-})
-
-test('matches request header with regular expression', t => {
-  nock('http://example.test', {
-    reqheaders: {
-      'X-My-Super-Power': /.+/,
-    },
-  })
-    .post('/superpowers')
-    .reply(200, { status: 'ok' })
-
-  mikealRequest(
-    {
-      method: 'POST',
-      uri: 'http://example.test/superpowers',
-      headers: {
-        'X-My-Super-Power': 'mullet growing',
-      },
-    },
-    function(err, res, body) {
-      t.strictEqual(err, null)
-      t.equal(body, '{"status":"ok"}')
-      t.end()
-    }
-  )
-})
-
-test('request header satisfies the header function', t => {
-  nock('http://example.test', {
-    reqheaders: {
-      'X-My-Super-Power': function(value) {
-        return value === 'mullet growing'
-      },
-    },
-  })
-    .post('/superpowers')
-    .reply(200, { status: 'ok' })
-
-  mikealRequest(
-    {
-      method: 'POST',
-      uri: 'http://example.test/superpowers',
-      headers: {
-        'X-My-Super-Power': 'mullet growing',
-      },
-    },
-    function(err, res, body) {
-      t.strictEqual(err, null)
-      t.equal(body, '{"status":"ok"}')
-      t.end()
-    }
-  )
-})
-
-test("done fails when specified request header doesn't match regular expression", t => {
-  nock('http://example.test', {
-    reqheaders: {
-      'X-My-Super-Power': /Mullet.+/,
-    },
-  })
-    .post('/resource')
-    .reply(200, { status: 'ok' })
-
-  const d = domain.create()
-
-  d.run(function() {
-    mikealRequest({
-      method: 'POST',
-      uri: 'http://example.test/superpowers',
-      headers: {
-        'X-My-Super-Power': 'mullet growing',
-      },
-    })
-  })
-
-  d.once('error', function(err) {
-    t.ok(err.message.match(/No match/))
-    t.end()
-  })
-})
-
-test("done fails when specified request header doesn't satisfy the header function", t => {
-  nock('http://example.test', {
-    reqheaders: {
-      'X-My-Super-Power': function(value) {
-        return value === 'Mullet Growing'
-      },
-    },
-  })
-    .post('/resource')
-    .reply(200, { status: 'ok' })
-
-  const d = domain.create()
-
-  d.run(function() {
-    mikealRequest({
-      method: 'POST',
-      uri: 'http://example.test/superpowers',
-      headers: {
-        'X-My-Super-Power': 'mullet growing',
-      },
-    })
-  })
-
-  d.once('error', function(err) {
-    t.ok(err.message.match(/No match/))
-    t.end()
-  })
-})
-
-test('done does not fail when specified request header is not missing', t => {
-  nock('http://example.test', {
-    reqheaders: {
-      'X-App-Token': 'apptoken',
-      'X-Auth-Token': 'apptoken',
-    },
-  })
-    .post('/resource')
-    .reply(200, { status: 'ok' })
-
-  mikealRequest(
-    {
-      method: 'POST',
-      uri: 'http://example.test/resource',
-      headers: {
-        'X-App-Token': 'apptoken',
-        'X-Auth-Token': 'apptoken',
-      },
-    },
-    function(err, res, body) {
-      t.type(err, 'null')
-      t.equal(res.statusCode, 200)
-      t.end()
-    }
-  )
-})
-
-test('done fails when specified bad request header is present', t => {
-  nock('http://example.test', {
-    badheaders: ['cookie'],
-  })
-    .post('/resource')
-    .reply(200, { status: 'ok' })
-
-  const d = domain.create()
-
-  d.run(function() {
-    mikealRequest({
-      method: 'POST',
-      uri: 'http://example.test/resource',
-      headers: {
-        Cookie: 'cookie',
-      },
-    })
-  })
-
-  d.once('error', function(err) {
-    t.ok(err.message.match(/No match/))
-    t.end()
-  })
-})
-
 test('get correct filtering with scope and request headers filtering', t => {
   const responseText = 'OK!'
   const responseHeaders = { 'Content-Type': 'text/plain' }
@@ -2476,59 +1266,6 @@ test('get correct filtering with scope and request headers filtering', t => {
   )
 
   t.equivalent(req._headers, { host: requestHeaders.host })
-})
-
-test('mocking succeeds even when mocked and specified request header names have different cases', t => {
-  nock('http://example.test', {
-    reqheaders: {
-      'x-app-token': 'apptoken',
-      'x-auth-token': 'apptoken',
-    },
-  })
-    .post('/resource')
-    .reply(200, { status: 'ok' })
-
-  mikealRequest(
-    {
-      method: 'POST',
-      uri: 'http://example.test/resource',
-      headers: {
-        'X-App-TOKEN': 'apptoken',
-        'X-Auth-TOKEN': 'apptoken',
-      },
-    },
-    function(err, res, body) {
-      t.type(err, 'null')
-      t.equal(res.statusCode, 200)
-      t.end()
-    }
-  )
-})
-
-// https://github.com/nock/nock/issues/966
-test('mocking succeeds when mocked and specified request headers have falsy values', t => {
-  nock('http://example.test', {
-    reqheaders: {
-      'x-foo': 0,
-    },
-  })
-    .post('/resource')
-    .reply(200, { status: 'ok' })
-
-  mikealRequest(
-    {
-      method: 'POST',
-      uri: 'http://example.test/resource',
-      headers: {
-        'X-Foo': 0,
-      },
-    },
-    function(err, res, body) {
-      t.error(err)
-      t.equal(res.statusCode, 200)
-      t.end()
-    }
-  )
 })
 
 test('mocking succeeds even when host request header is not specified', t => {
@@ -2596,43 +1333,6 @@ test('.setNoDelay', t => {
   req.setNoDelay(true)
 
   req.end()
-})
-
-test('match basic authentication header', t => {
-  const username = 'testuser'
-  const password = 'testpassword'
-  const authString = `${username}:${password}`
-
-  const expectedAuthHeader = `Basic ${Buffer.from(authString).toString(
-    'base64'
-  )}`
-
-  const scope = nock('http://example.test')
-    .get('/')
-    .matchHeader('Authorization', val => val === expectedAuthHeader)
-    .reply(200, 'Hello World!')
-
-  http.get(
-    {
-      host: 'example.test',
-      path: '/',
-      port: 80,
-      auth: authString,
-    },
-    function(res) {
-      res.setEncoding('utf8')
-      t.equal(res.statusCode, 200)
-
-      res.on('data', function(data) {
-        t.equal(data, 'Hello World!')
-      })
-
-      res.on('end', function() {
-        scope.done()
-        t.end()
-      })
-    }
-  )
 })
 
 test('request emits socket', t => {
@@ -2891,41 +1591,6 @@ test('match path using function', t => {
     .end()
 })
 
-test('isDone() must consider repeated responses', t => {
-  const scope = nock('http://example.test')
-    .get('/')
-    .times(2)
-    .reply(204)
-
-  function makeRequest(callback) {
-    const req = http.request(
-      {
-        host: 'example.test',
-        path: '/',
-        port: 80,
-      },
-      function(res) {
-        t.equal(res.statusCode, 204)
-        res.on('end', callback)
-        // Streams start in 'paused' mode and must be started.
-        // See https://nodejs.org/api/stream.html#stream_class_stream_readable
-        res.resume()
-      }
-    )
-    req.end()
-  }
-
-  t.notOk(scope.isDone(), 'should not be done before all requests')
-  makeRequest(function() {
-    t.notOk(scope.isDone(), 'should not yet be done after the first request')
-    makeRequest(function() {
-      t.ok(scope.isDone(), 'should be done after the two requests are made')
-      scope.done()
-      t.end()
-    })
-  })
-})
-
 test('you must setup an interceptor for each request', t => {
   const scope = nock('http://example.test')
     .get('/hey')
@@ -2952,51 +1617,6 @@ test('you must setup an interceptor for each request', t => {
       t.end()
     })
   })
-})
-
-test('replyWithError returns an error on request', t => {
-  const scope = nock('http://example.test')
-    .post('/echo')
-    .replyWithError('Service not found')
-
-  const req = http.request({
-    host: 'example.test',
-    method: 'POST',
-    path: '/echo',
-    port: 80,
-  })
-
-  // An error should have have been raised
-  req.on('error', function(e) {
-    scope.done()
-    t.equal(e.message, 'Service not found')
-    t.end()
-  })
-
-  req.end()
-})
-
-test('replyWithError allows json response', t => {
-  const scope = nock('http://example.test')
-    .post('/echo')
-    .replyWithError({ message: 'Service not found', code: 'test' })
-
-  const req = http.request({
-    host: 'example.test',
-    method: 'POST',
-    path: '/echo',
-    port: 80,
-  })
-
-  // An error should have have been raised
-  req.on('error', function(e) {
-    scope.done()
-    t.equal(e.message, 'Service not found')
-    t.equal(e.code, 'test')
-    t.end()
-  })
-
-  req.end()
 })
 
 test('no content type provided', t => {
@@ -3049,60 +1669,6 @@ test('match domain and path using regexp', t => {
   })
 })
 
-test('multiple interceptors override headers from unrelated request', t => {
-  nock.cleanAll()
-
-  nock.define([
-    {
-      scope: 'https://example.test:443',
-      method: 'get',
-      path: '/bar',
-      reqheaders: {
-        'x-foo': 'bar',
-      },
-      status: 200,
-      response: {},
-    },
-    {
-      scope: 'https://example.test:443',
-      method: 'get',
-      path: '/baz',
-      reqheaders: {
-        'x-foo': 'baz',
-      },
-      status: 200,
-      response: {},
-    },
-  ])
-
-  mikealRequest(
-    {
-      url: 'https://example.test/bar',
-      headers: {
-        'x-foo': 'bar',
-      },
-    },
-    function(err, res, body) {
-      t.error(err)
-      t.equal(res.statusCode, 200)
-
-      mikealRequest.get(
-        {
-          url: 'https://example.test/baz',
-          headers: {
-            'x-foo': 'baz',
-          },
-        },
-        function(err, res, body) {
-          t.error(err)
-          t.equal(res.statusCode, 200)
-          t.end()
-        }
-      )
-    }
-  )
-})
-
 // https://github.com/nock/nock/issues/1003
 test('correctly parse request without specified path', t => {
   nock.cleanAll()
@@ -3142,23 +1708,6 @@ test('data is sent with flushHeaders', t => {
       })
     })
     .flushHeaders()
-})
-
-test('should log matching', async t => {
-  const messages = []
-
-  const scope = nock('http://example.test')
-    .get('/')
-    .reply(200, 'Hello, World!')
-    .log(message => messages.push(message))
-
-  await got('http://example.test/')
-
-  t.deepEqual(messages, [
-    'matching http://example.test:80/ to GET http://example.test:80/: true',
-  ])
-
-  scope.done()
 })
 
 test('teardown', t => {
